add authentication

+++ twitter-clone/app/Http/Controllers/IdeaController.php
@@ -17,6 +17,7 @@ class IdeaController extends Controller
             'content' => 'required|min:5|max:300'
         ]);
 
+        $validated['user_id'] = auth()->id();
         Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success','Idea Created Successfully');  
@@ -24,26 +25,30 @@ class IdeaController extends Controller
     }
 
     public function destroy(Idea $idea){
+       if (auth()->id() !== $idea->user_id){
+            abort(404);
+       }  
        $idea->delete();
 
        return redirect()->route('dashboard')->with('success','Idea deleted Successfully');  
     }
 
-    // public function destroy($id){
-    //     Idea::where('id',$id)->firstOrFail()->delete();
-    // //    $idea->delete();
 
-    //    return redirect()->route('dashboard')->with('success','Idea deleted Successfully');  
-    // }
 
     
 
     public function edit(Idea $idea){
+        if (auth()->id() !== $idea->user_id){
+            abort(404);
+        } 
         $editing = true;
         return view('ideas.show',compact('idea', 'editing'));
     }
 
     function update(Idea $idea){
+        if (auth()->id() !== $idea->user_id){
+            abort(404);
+        } 
         $validated = request()->validate([
             'content' => 'required|min:5|max:300'
         ]);
